+++ packages/plugins/other/visitor-plugin-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @graphql-codegen/visitor-plugin-common
 
+## 6.1.0
+
+### Minor Changes
+
+- [#10449](https://github.com/dotansimha/graphql-code-generator/pull/10449) [`8258f1f`](https://github.com/dotansimha/graphql-code-generator/commit/8258f1f6012c106d02ef28bca9ec424f70c4aa26) Thanks [@eddeee888](https://github.com/eddeee888)! - Add addInterfaceFieldResolverTypes option to support custom Interface resolver inheritance
+
 ## 6.0.1
 
 ### Patch Changes