+++ packages/plugins/typescript/typed-document-node/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @graphql-codegen/typed-document-node
 
+## 6.0.2
+
+### Patch Changes
+
+- Updated dependencies [[`8258f1f`](https://github.com/dotansimha/graphql-code-generator/commit/8258f1f6012c106d02ef28bca9ec424f70c4aa26)]:
+  - @graphql-codegen/visitor-plugin-common@6.1.0
+
 ## 6.0.1
 
 ### Patch Changes